requirements.txt
yfinance
pandas
requests
google-generativeai